requirements.txt
requests
feedparser
youtube-transcript-api
openai
python-dotenv
pytz
